Category filters issue solved
Problem: At the beginning we had all names of category in the select
field ("multiple" type of field). Those names didn't fit in the
selection field and went out of the field. In this case the "selection
field" should shows only number of selected categories.

Recognition: When the site is loaded for the first time advanced filters
dosen't exist (the "mr-filters-advanced" class sets the display argument
to "none"). Because of that the jQuery-MultiSelect component dosne't
work properly.
Two action have influance fot the showing and hiding advanced menu. It
is:
1. "click" event, which is situated in sample.js file.
2. "showFilters" function, situated in the "mainPageController".

Solution:
1. Removed click event from sample.js file, which overwrote the action
from the "showFilters" function.
2. Changed default displaing of the advanced menu to "block"
("mr-filters-advanced" class).